Add function to put extracted text
To put extracted text, we use opencv and according to the bounding box coordinates we put the text over image.

At the moment, due to absence of a working program, I have not calibrated the function to put text size according to box but it'll be done once we have other parts working.